replication: introduce orphan mode
First of all, it's worth mentioning that we already have 'orphan' status
in Tarantool (between 'loading' and 'hot_standby'), but it has nothing
to do with replication. Actually, it's unclear why it was introduced in
the first place so we agreed to drop it. Now, to the new 'orphan' mode.

Tarantool is switched to 'orphan' mode if it failed to form a quorum
within the specified timeout during initial configuration of an instance
that has already been bootstrapped. Basically, 'orphan' status forces
the instance to switch to read-only mode. The instance will switch back
to 'running' (and enable write access) in either of the following three
cases:

 - replication is successfully reconfigured
 - box.cfg.replication_connect_quorum is changed
 - enough replicas are connected to form a quorum

Note, any subsequent attempt to change box.cfg.replication won't change
the server status - as before, it will fail downright if it can't form a
quorum within the configured timeout period.

Just like any other status, the new one is reflected by box.info.status.

Suggested by @kostja

Follow-up #2958